Fix MPS float64 crash in ONNX-test and reference training scripts
The ONNX evaluation scripts (references/*/test_onnx.py) and the anomaly-
detection training script accumulated results with `torch.tensor([])`, whose
default dtype is float64, and moved batches to the device before casting
(`.to(device).float()` / `.to(device).long()`). On Apple-Silicon MPS both
patterns raise "Cannot convert a MPS Tensor to float64 dtype".

- Pin the accumulator tensors to `dtype=torch.float32`.
- Cast (`.float()`/`.long()`) before `.to(device)` so the transfer never
  carries float64.

Covers timeseries (classification, regression, forecasting, anomaly-
detection) plus audio and image classification. No-op on CUDA/CPU.

(cherry picked from commit 9bf2289fd38c288540b8684a9c603cd13c7062ae)

+++ tinyml-tinyverse/tinyml_tinyverse/references/image_classification/test_onnx.py
@@ -135,12 +135,12 @@ def main(gpu, args):
     logger.info(f"Loading ONNX model: {args.model_path}")
     ort_sess, input_name, output_name = load_onnx_model(args.model_path, args.generic_model)
 
-    predicted = torch.tensor([]).to(device, non_blocking=True)
-    ground_truth = torch.tensor([]).to(device, non_blocking=True)
+    predicted = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
+    ground_truth = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
     for batched_raw_data, batched_data, batched_target in data_loader:
-        batched_raw_data = batched_raw_data.to(device, non_blocking=True).long()
-        batched_data = batched_data.to(device, non_blocking=True).float()
-        batched_target = batched_target.to(device, non_blocking=True).long()
+        batched_raw_data = batched_raw_data.long().to(device, non_blocking=True)
+        batched_data = batched_data.float().to(device, non_blocking=True)
+        batched_target = batched_target.long().to(device, non_blocking=True)
         if transform:
             batched_data = transform(batched_data)
         if args.nn_for_feature_extraction:
